registration controller wip
