+++ src/controllers/plane_fitting_controller.py
@@ -1,8 +1,10 @@
 from PySide6.QtCore import Signal
-from sympy import Plane
 
 from controllers.base_controller import BaseController
+from gui.widgets.progress_dialog_factory import ProgressDialogFactory
 from gui.workers.downsampling.qt_plane_fitting import PlaneFittingWorker
+from gui.workers.downsampling.qt_plane_merging import PlaneInlierMergingWorker
+from gui.workers.qt_base_worker import move_worker_to_thread
 from models.data_repository import DataRepository
 from params.plane_fitting_params import PlaneFittingParams
 from utils.plane_fitting_util import get_o3d_plane
@@ -16,6 +18,7 @@ class PlaneFittingController(BaseController):
 
     # region Event handlers
     def fit_plane(self, params: PlaneFittingParams):
+        # TODO: Actually we should consider the point cloud after the transformation
         pc1 = self.repository.pc_open3d_list_first[0]
         pc2 = self.repository.pc_open3d_list_second[0]
 
@@ -30,7 +33,6 @@ class PlaneFittingController(BaseController):
         self.repository.second_plane_indices.clear()
         self.repository.first_plane_coefficients.clear()
         self.repository.second_plane_coefficients.clear()
-
     # endregion
 
     # region Result handlers
@@ -55,5 +57,5 @@ class PlaneFittingController(BaseController):
             planes.append(get_o3d_plane(result_data.coefficients_pc2[i], pc2.get_xyz[result_data.indices_pc2[i]],
                                         [0.8, 0.1, 0.1]))
 
-        self.signal_add_planes.emit(planes)
+        self.signal_add_planes.emit(planes)  # FIXME: Maybe store these in the repository as well? --> Trigger UI update
     # endregion
